pyproject.toml
[build-system]
# 声明使用 setuptools 作为打包后端
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "onebase-ai"  # 在 PyPI 上的包名，建议加个后缀防止与已有包冲突
version = "0.1.3"  # 遵循语义化版本控制
description = "像配置静态网站一样，一键构建与部署 AI 动态服务 (RAG 底座)"
readme = "README.md"
authors = [
    { name = "pirate-608", email = "huangmaidou608@gmail.com" }
]
requires-python = ">=3.10"
# 将 onebase/requirements.txt 中的核心依赖搬到这里
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "langchain-core>=0.2.0",        # LangChain 核心抽象类
    "langchain-postgres>=0.0.9",    # PGVector 入库必备
    "psycopg[binary]>=3.0.0",       # 数据库驱动
    "python-dotenv>=1.0.0",
    # 兼容性锁定
    "requests>=2.32.0,<=2.32.4",
    "tenacity>=8.5.0,<10.0.0",
    "pyjwt[crypto]>=2.10.1"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "fastapi>=0.100",
    "httpx>=0.24",
    "python-multipart>=0.0.5",
]

# 🚀 核心魔法：将代码注册为系统命令
[project.scripts]
# 当用户在终端输入 onebase 时，实际执行的是 onebase.cli 模块下的 app 对象
onebase = "onebase.cli:app"

[tool.setuptools]
# 告诉打包工具，我们的源码在哪些目录
# 必须包含 templates 目录，因为 onebase CLI 需要将它们拷贝给用户
packages = ["onebase", "templates"]

[tool.setuptools.package-data]
# 确保打包时不会漏掉静态资源、Dockerfile、依赖清单等非 Python 文件
"templates" = ["**/*", "**/.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"